Update the standard settings file
So that it's less opinionated about things that are repository specific,
like if something has a wiki. And more opinionated about things like
which teams have access to the repository and which branches are
protected. This will enable us to have standardised permissions across
all our repositories so anyone in our group can maintain them.