Add stricter protections on the main branches
Don't allow branches to be merged to main until a PR has been reviewed
and tests have passed. Also make explicit that force pushes to main and
deleting main are disallowed, although this is already the default
behaviour.

These changes should hopefully prevent accidental commits to main and
make it clearer when it is safe to merge code into the main branch.